Refactor how old event configs are stored
